+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.x branch
 ## 1.0 branch
 ### 1.0.6
+* added `rmdir_recursive()` and `safe_rmdir_recursive()` functions;
 * added `file_exists_or_fail()`, `is_readable_or_fail()`, `is_writable_or_fail()`
     functions;
 * added `assertIsArrayNotEmpty()` assertion method.